Make local port conflict detector recognize varying target hosts
The Telepresence daemon's local port conflict detection assumed that the
localhost was used for all client ports targeted by an intercept. In
reality, and when using `telepresence compose` specifically, there can
be multiple target hosts. The conflict detection now takes this into
account so that the same port can be used on different hosts.

Signed-off-by: Thomas Hallgren <thomas@tada.se>